fix: [#S10P12A402-98] edit Long type to Integer type

+++ backend/src/main/java/org/ssafy/pasila/domain/product/entity/ProductOption.java
@@ -20,9 +20,9 @@ public class ProductOption {
     @Column(length = 20)
     private String name;
 
-    private Long stock;
+    private Integer stock;
 
-    private Long price;
+    private Integer price;
 
     @Column(name = "discount_price")
     private Long discountPrice;
